feat: Refactor repo, add optical flow
This commit refactors the repo to better represent what will be needed to extract
segmented objects from videos in an effort to make it more streamlined. There
is still some work that needs to be done but this is a big chunk of it. This is
also to make it more my own repo based on the original instead of just making
tweaks to the original.

I also wrote an optical flow function because I want to test if using optical
flow will improve the automatic image segmentation using the turntable. It
still needs to be tested but it is the Lucas-Kanade method and is pretty much
just copied from a few Matlab files I wrote for class last year.