Shift repo focus to Mask-RCNN (#1)
* now outputs masks

* indexing added

* added rotation and attempt at occlusion fix

* attempt to fix occlusion errors

* added multithreading

* perfected multithreading

* cleaned up darknet code

* added multi and input dir path

* added better README

* added background step

* added jpg extension comment

* Update README.md

* auto png generator

* perfected multithreading

* feat: Generate composites successful

`vids2pngs.py`, `auto_generate_png.py` and `generate_composites_multithread.py`
are all working, though the PNGs are not the best quality so I need to look at
that. Next step is to create the training script (WIP) and use it to train the
model for slippers.

* feat: Add improved background subtraction methods

The updated methods average the background images to create a clean background
template, then subtract the object images from the background to isolate the
objects. Still needs a little work but it is working well already.

* chore: Update progress message

* feat: Refactor repo, add optical flow

This commit refactors the repo to better represent what will be needed to extract
segmented objects from videos in an effort to make it more streamlined. There
is still some work that needs to be done but this is a big chunk of it. This is
also to make it more my own repo based on the original instead of just making
tweaks to the original.

I also wrote an optical flow function because I want to test if using optical
flow will improve the automatic image segmentation using the turntable. It
still needs to be tested but it is the Lucas-Kanade method and is pretty much
just copied from a few Matlab files I wrote for class last year.

* refactor: Clean up repo

* refactor: Clean up repo and improve docs

* docs: Fix README spacing

* docs: Fix README spacing

* docs: Fix README spacing